Update README.md
IBM's MASTER of Data Science Professional Certificate.
• What is Data Science? • Open Source tools for Data Science • Data Science Methodology • Python for Data Science and AI • Databases and SQL for Data Science • Data Analysis with Python • Data Visualization with Python • Machine Learning with Python • Applied Data Science Capstone
In this Professional Certificate learners developed and honed hands-on skills in Data Science and Machine Learning. Sergei Fotin started with an orientation of Data Science and its Methodology, became familiar and used a variety of data science tools, learned Python and SQL, performed Data Visualization and Analysis, and created Machine Learning models. In the process they completed several labs and assignments on the cloud including a Capstone Project at the end to apply and demonstrate their knowledge and skills.